code/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
google-generativeai==0.3.2
numpy==1.24.3
pydantic-settings==2.1.0
python-multipart==0.0.6
